modify for sys---not finished

+++ run_MC/job_combine.sh
@@ -15,7 +15,7 @@ do
 done
 
 #test the number of parameter
-if (($#!=2));then
+if (($#!=1));then
 echo "ERROR:the number of parameter is invalid!"
 echo "usage:systematics? 0 / 1 / 2 -- nom / nom+sys / sys"
 exit
@@ -24,7 +24,7 @@ fi
 DO_SYS=${1}
 
 #syslist
-syslist=("JET_EffectiveNP_R10_6restTerm__1up" "JET_EffectiveNP_R10_6restTerm__1down" "JET_EtaIntercalibration_Modelling__1up" "JET_EtaIntercalibration_Modelling__1down" "JET_EtaIntercalibration_NonClosure_2018data__1up" "JET_EtaIntercalibration_NonClosure_2018data__1down" "JET_EtaIntercalibration_R10_TotalStat__1up" "JET_EtaIntercalibration_R10_TotalStat__1down" "JET_EffectiveNP_R10_1__1up" "JET_EffectiveNP_R10_1__1down" "JET_EffectiveNP_R10_2__1up" "JET_EffectiveNP_R10_2__1down" "JET_EffectiveNP_R10_3__1up" "JET_EffectiveNP_R10_3__1down" "JET_EffectiveNP_R10_4__1up" "JET_EffectiveNP_R10_4__1down" "JET_EffectiveNP_R10_5__1up" "JET_EffectiveNP_R10_5__1down" "JET_Flavor_Composition__1up" "JET_Flavor_Composition__1down" "JET_Flavor_Response__1up" "JET_Flavor_Response__1down" "MUON_ID__1up" "MUON_ID__1down" "MUON_SAGITTA_RESBIAS__1up" "MUON_SAGITTA_RESBIAS__1down" "MUON_SCALE__1up" "MUON_SCALE__1down" "MUON_EFF_TTVA_SYS__1up" "MUON_EFF_TTVA_SYS__1down" "TRK_RES_D0_MEAS__1up" "TRK_RES_D0_MEAS__1down" "TRK_RES_Z0_MEAS__1up" "TRK_RES_Z0_MEAS__1down" "TRK_RES_D0_DEAD" "TRK_RES_Z0_DEAD" "TRK_BIAS_D0_WM" "TRK_BIAS_QOVERP_SAGITTA_WM" "TRK_BIAS_Z0_WM" "TRK_EFF_LOOSE_GLOBAL" "TRK_EFF_LOOSE_IBL" "TRK_EFF_LOOSE_PHYSMODEL" "TRK_EFF_LOOSE_PP0" "TRK_FAKE_RATE_LOOSE" "TRK_EFF_LOOSE_TIDE" "TRK_FAKE_RATE_LOOSE_TIDE")
+syslist=("JET_EffectiveNP_R10_6restTerm__1up" "JET_EffectiveNP_R10_6restTerm__1down" "JET_EtaIntercalibration_Modelling__1up" "JET_EtaIntercalibration_Modelling__1down" "JET_EtaIntercalibration_NonClosure_2018data__1up" "JET_EtaIntercalibration_NonClosure_2018data__1down" "JET_EtaIntercalibration_R10_TotalStat__1up" "JET_EtaIntercalibration_R10_TotalStat__1down" "JET_EffectiveNP_R10_1__1up" "JET_EffectiveNP_R10_1__1down" "JET_EffectiveNP_R10_2__1up" "JET_EffectiveNP_R10_2__1down" "JET_EffectiveNP_R10_3__1up" "JET_EffectiveNP_R10_3__1down" "JET_EffectiveNP_R10_4__1up" "JET_EffectiveNP_R10_4__1down" "JET_EffectiveNP_R10_5__1up" "JET_EffectiveNP_R10_5__1down" "JET_Flavor_Composition__1up" "JET_Flavor_Composition__1down" "JET_Flavor_Response__1up" "JET_Flavor_Response__1down" "MUON_ID__1up" "MUON_ID__1down" "MUON_SAGITTA_RESBIAS__1up" "MUON_SAGITTA_RESBIAS__1down" "MUON_SCALE__1up" "MUON_SCALE__1down" "MUON_EFF_TTVA_SYS__1up" "MUON_EFF_TTVA_SYS__1down" "TRK_RES_D0_MEAS__1up" "TRK_RES_D0_MEAS__1down" "TRK_RES_Z0_MEAS__1up" "TRK_RES_Z0_MEAS__1down" "TRK_RES_D0_DEAD" "TRK_RES_Z0_DEAD" "TRK_EFF_LOOSE_GLOBAL" "TRK_EFF_LOOSE_IBL" "TRK_EFF_LOOSE_PHYSMODEL" "TRK_EFF_LOOSE_PP0" "TRK_FAKE_RATE_LOOSE" "TRK_EFF_LOOSE_TIDE" "TRK_FAKE_RATE_LOOSE_TIDE")
 
 # systematics explanation
 if [ "${DO_SYS}" -eq 0 ]
@@ -59,6 +59,13 @@ do
  RootName=(`ls`)
  RootNum=`ls | wc -l`
 
+ echo $path"/"$file
+ echo $RootNum
+ if (($RootNum==0));then #no file in this directory
+  echo >&1000
+  continue
+ fi
+
  cal_list=0
  cal_file=0
 
@@ -86,7 +93,10 @@ do
  done
 
  #add whole root file
- rm outhist_$file.root
+ if [ -f "outhist_$file.root" ];then
+  echo "hello!"
+  rm outhist_$file.root
+ fi
  AddWhole="hadd outhist_$file.root"
  for ((i=1;i<=cal_list;i++))
  do
